src/requirements.txt
pandas
minio
psycopg
python-dotenv
requests
boto3
pyarrow
fastparquet